+++ Makefile.am
@@ -134,7 +134,7 @@ src_libfabric_la_SOURCES =			\
 	include/fi_proto.h			\
 	include/fi_rbuf.h			\
 	include/fi_shm.h			\
-	include/fi_signal.h			\
+	include/ofi_signal.h			\
 	include/fi_util.h			\
 	include/ofi_atomic.h			\
 	include/ofi_mr.h			\
