+++ libfabric.vcxproj
@@ -581,7 +581,7 @@
     <ClInclude Include="include\ofi_perf.h" />
     <ClInclude Include="include\fi_proto.h" />
     <ClInclude Include="include\fi_rbuf.h" />
-    <ClInclude Include="include\fi_signal.h" />
+    <ClInclude Include="include\ofi_signal.h" />
     <ClInclude Include="include\fi_util.h" />
     <ClInclude Include="include\prov.h" />
     <ClInclude Include="include\rbtree.h" />
