+++ prov/shm/src/smr.h
@@ -56,7 +56,7 @@
 #include <fi_shm.h>
 #include <fi_rbuf.h>
 #include <fi_list.h>
-#include <fi_signal.h>
+#include <ofi_signal.h>
 #include <fi_util.h>
 #include <ofi_atomic.h>
