+++ prov/udp/src/udpx.h
@@ -57,7 +57,7 @@
 #include <fi_enosys.h>
 #include <fi_rbuf.h>
 #include <fi_list.h>
-#include <fi_signal.h>
+#include <ofi_signal.h>
 #include <fi_util.h>
 
 #ifndef _UDPX_H_
