+++ src/common.c
@@ -50,7 +50,7 @@
 #include <netinet/in.h>
 #include <arpa/inet.h>
 
-#include <fi_signal.h>
+#include <ofi_signal.h>
 #include <rdma/providers/fi_prov.h>
 #include <rdma/fi_errno.h>
 #include <fi.h>
